requirements.txt
fastapi
uvicorn
requests
python-dateutil